Revert "Halfs"
This reverts commit 6526b72c20c4da42f2d10e170bdb45184c93d9fb.

+++ HKX2/Manual/hkaiDirectedGraphExplicitCostEdge.cs
@@ -2,21 +2,21 @@ namespace HKX2
 {
     public class hkaiDirectedGraphExplicitCostEdge : IHavokObject
     {
-        public System.Half m_cost;
+        public short m_cost;
         public EdgeBits m_flags;
         public uint m_target;
         public virtual uint Signature => 0;
 
         public virtual void Read(PackFileDeserializer des, BinaryReaderEx br)
         {
-            m_cost = br.ReadHalf();
+            m_cost = br.ReadInt16();
             m_flags = (EdgeBits) br.ReadUInt16();
             m_target = br.ReadUInt32();
         }
 
         public virtual void Write(PackFileSerializer s, BinaryWriterEx bw)
         {
-            bw.WriteHalf(m_cost);
+            bw.WriteInt16(m_cost);
             bw.WriteUInt16((ushort) m_flags);
             bw.WriteUInt32(m_target);
         }
